+++ packages/cdk8s-cli/test/import/type-generator.test.ts
@@ -17,6 +17,59 @@ describe('unions', () => {
     ]
   });
 
+  which('constraints are ignored for objects', {
+    "description": "An ordered list of route rules for HTTP traffic.",
+    "type": "array",
+    "items": {
+      "type": "object",
+      "properties": {
+        "fault": {
+          "type": "object",
+          "description": "Fault injection policy to apply on HTTP traffic at\nthe client side.",
+          "properties": {
+            "delay": {
+              "oneOf": [
+                {
+                  "anyOf": [
+                    { "required": [ "fixedDelay" ] },
+                    { "required": [ "exponentialDelay" ] }
+                  ]
+                },
+                { "required": [ "fixedDelay" ] },
+                { "required": [ "exponentialDelay" ] }
+              ],
+              "properties": {
+                "exponentialDelay": {
+                  "type": "string"
+                },
+                "fixedDelay": {
+                  "description": "Add a fixed delay before forwarding the request.",
+                  "type": "string"
+                },
+                "percent": {
+                  "description": "Percentage of requests on which the delay\nwill be injected (0-100).",
+                  "format": "int32",
+                  "type": "integer"
+                },
+                "percentage": {
+                  "description": "Percentage of requests on which the delay\nwill be injected.",
+                  "properties": {
+                    "value": {
+                      "format": "double",
+                      "type": "number"
+                    }
+                  },
+                  "type": "object"
+                }
+              },
+              "type": "object"
+            }
+          },
+        }
+      },
+    },
+  });
+
 });
 
 
@@ -140,17 +193,58 @@ describe('structs', () => {
 describe('documentation', () => {
 
   which('does not render if not defined', {
-    type: 'boolean'
+    type: 'object',
+    properties: {
+      field: {
+        type: 'boolean'
+      }
+    }
   });
 
   which('renders based on description', {
-    description: 'hello, description',
-    type: 'string'
+    type: 'object',
+    properties: {
+      field: {
+        description: 'hello, description',
+        type: 'string'
+      }
+    }
   });
 
   which('"*/" is is escaped', {
-    description: 'hello */world',
-    type: 'string'
+    type: 'object',
+    properties: {
+      field: {
+        description: 'hello */world',
+        type: 'string'
+      }
+    }
+  });
+
+});
+
+describe('enums', () => {
+
+  which('renders a typescript enum', {
+    type: 'object',
+    required: [ 'firstEnum' ],
+    properties: {
+      firstEnum: {
+        description: 'description of first enum',
+        type: 'string',
+        enum: [ 'value1', 'value2', 'value-of-three', 'valueOfFour' ]
+      },
+      child: {
+        type: 'object',
+        properties: {
+          secondEnum: {
+            description: 'description of second enum',
+            type: 'string',
+            enum: [ 'hey', 'enum values can be crazy', 'yes>>123' ]
+          }
+        }
+      }
+    }
   });
 
 });